[2.0.14] 일반 이미지 업로드시 960 x 540 으로 리사이즈 + 이전 파일 삭제기능 개선
- 이전 파일의 확장자가 png 가 아닌 경우 not found 에러발생하므로 이와 같이 수정

+++ src/main/mainUtils.ts
@@ -8,7 +8,7 @@ import { performance } from 'perf_hooks';
 import root from 'window-or-global';
 import Puid from 'puid';
 import uid from 'uid';
-import FileUtils from './fileUtils';
+import FileUtils, { ImageResizeSize } from './fileUtils';
 import Constants, { ReplaceStrategy } from './constants';
 import CommonUtils from './commonUtils';
 import BlockConverter from './blockConverter';
@@ -412,14 +412,17 @@ export default class MainUtils {
         const newPicturePath = path.join(Constants.tempImagePath(newFileId), newFileName);
         const newThumbnailPath = path.join(Constants.tempThumbnailPath(newFileId), newFileName);
 
-        await FileUtils.copyFile(filePath, newPicturePath);
+        await FileUtils.writeFile(
+            FileUtils.createResizedImageBuffer(filePath, ImageResizeSize.picture),
+            newPicturePath,
+        );
 
         // 섬네일 이미지가 이미 있는 경우는 해당 이미지를 가져다 쓰고, 없는 경우 원본을 리사이징
         if (thumbnailPath) {
             await FileUtils.copyFile(thumbnailPath, newThumbnailPath);
         } else {
             await FileUtils.writeFile(
-                FileUtils.createThumbnailBuffer(filePath),
+                FileUtils.createResizedImageBuffer(filePath, ImageResizeSize.thumbnail),
                 newThumbnailPath,
             );
         }
@@ -469,19 +472,23 @@ export default class MainUtils {
                 await Promise.all([
                     FileUtils.writeFile(image, imagePath),
                     FileUtils.writeFile(
-                        FileUtils.createThumbnailBuffer(image),
+                        FileUtils.createResizedImageBuffer(image, ImageResizeSize.thumbnail),
                         thumbnailPath),
                 ]);
 
                 // 편집모드이며 리소스 기본이미지가이 아닌, temp 내 원본 이미지가 있는 경우 이전 이미지를 삭제한다.
                 if (prevFilename && mode === 'edit') {
-                    const prevImagePath = path.join(Constants.tempImagePath(prevFilename), `${prevFilename}.png`);
-                    const prevThumbnailPath = path.join(Constants.tempThumbnailPath(prevFilename), `${prevFilename}.png`);
+                    const prevImageFileName = fs.readdirSync(Constants.tempImagePath(prevFilename))[0];
 
-                    await Promise.all([
-                        FileUtils.deleteFile(prevImagePath),
-                        FileUtils.deleteFile(prevThumbnailPath),
-                    ]);
+                    if (prevImageFileName) {
+                        const prevImagePath = path.join(Constants.tempImagePath(prevImageFileName), prevImageFileName);
+                        const prevThumbnailPath = path.join(Constants.tempThumbnailPath(prevImageFileName), prevImageFileName);
+
+                        await Promise.all([
+                            prevImagePath && FileUtils.deleteFile(prevImagePath),
+                            prevThumbnailPath && FileUtils.deleteFile(prevThumbnailPath),
+                        ]);
+                    }
                 }
 
                 //TODO 빈 폴더인지 검사한 후, 삭제하기 (앞 4자리가 같은 다른 파일이 있을 수 있음)
@@ -518,7 +525,7 @@ export default class MainUtils {
 
         await FileUtils.copyFile(filePath, newSoundPath);
 
-        const metadata = await musicMetadata.parseFile(newSoundPath, {duration: true});
+        const metadata = await musicMetadata.parseFile(newSoundPath, { duration: true });
 
         return {
             _id: CommonUtils.generateHash(),
